Move error-location tests out of console-execution.spec.ts
Extract the "Check errors show" describe block into its own
check-error-locations.spec.ts, since it tests where errors get
reported/highlighted rather than console output behaviour in general.
checkFrameErrorCount is used throughout console-execution.spec.ts too,
so it moves to the shared support/execution.ts instead of being
re-exported from one spec file into another.
